Move test header to top level
