+++ protocol/clproto_cpp/test/tests/include/encode_decode_helper.hpp
