+++ protocol/clproto_cpp/test/tests/test_joint_space_proto.cpp
@@ -8,7 +8,7 @@
 #include <state_representation/space/joint/JointTorques.hpp>
 
 #include "clproto.h"
-#include "test_clproto/encode_decode_helper.hpp"
+#include "encode_decode_helper.hpp"
 
 using namespace state_representation;
