+++ protocol/clproto_cpp/test/tests/test_messages.cpp
@@ -6,7 +6,7 @@
 #include <state_representation/space/cartesian/CartesianPose.hpp>
 
 #include "clproto.h"
-#include "test_clproto/encode_decode_helper.hpp"
+#include "encode_decode_helper.hpp"
 
 using namespace state_representation;
